유효성 검사 및 버그 수정

+++ src/main/java/com/finalteam5/otisrm/dto/SrTrnsfPlan.java
@@ -15,6 +15,6 @@ public class SrTrnsfPlan {
 	private String instNo;			
 	private String srTrnsfNote;		
 	private String srPrgrsSttsNo;
-	private Integer totalCapacity;
+	private Float totalCapacity;
 	private Integer totalBusinessDt;
 }
